moving of bomber

+++ src/uet/oop/bomberman/entities/Bomber.java
@@ -9,7 +9,7 @@ import javafx.scene.paint.Color;
 
 public class Bomber extends Entity {
 
-    public Bomber(int x, int y, Image img) {
+    public Bomber(double x, double y, Image img) {
         super( x, y, img);
     }
